+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
+## [0.2.5] - 2026-02-02
+### Added
+- Slack: in "mention" mode, thread replies are forwarded if the bot has previously responded in that thread (`active_thread_ttl` configurable, default 24h)
+- Add permission debug to help troubleshoot Slack permission issues
+
 ## [0.2.4] - 2026-02-02
 ### Added
 - Publish workflow now blocks on security scans
@@ -82,6 +87,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Example files for all providers and relay components
 - Test documentation with coverage guidelines
 
+[0.2.5]: https://github.com/AgentSanchez/omni-message-provider/releases/tag/v0.2.5
 [0.2.2]: https://github.com/AgentSanchez/omni-message-provider/releases/tag/v0.2.2
 [0.2.3]: https://github.com/AgentSanchez/omni-message-provider/releases/tag/v0.2.3
 [0.2.4]: https://github.com/AgentSanchez/omni-message-provider/releases/tag/v0.2.4